+++ api/src/main/java/org/allaymc/api/entity/component/player/EntityPlayerBossBarViewerComponent.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.entity.component.player;
+
+import org.allaymc.api.bossbar.BossBarViewer;
+import org.allaymc.api.entity.component.EntityComponent;
+
+/**
+ * @author daoge_cmd
+ */
+public interface EntityPlayerBossBarViewerComponent extends EntityComponent, BossBarViewer {
+}